ERP project finished

+++ erp/src/main/java/com/allianz/erp/repository/BaseRepository.java
@@ -3,6 +3,6 @@ package com.allianz.erp.repository;
 import com.allianz.erp.entity.abstracts.BaseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BaseRepository<T extends BaseEntity> extends JpaRepository<T,Long> {
+public interface BaseRepository<T extends BaseEntity> extends JpaRepository<T,Integer> {
 
 }
